requirements.txt
# Local transcription dependencies
# Install with: pip install -r requirements.txt

# Core Whisper dependencies
openai-whisper>=20231117
torch>=2.1.0
torchaudio>=2.1.0

# Optional: WhisperX for better accuracy and speaker diarization
# whisperx>=3.1.1

# Audio processing
librosa>=0.10.1
soundfile>=0.12.1

# General dependencies
numpy>=1.24.0